Fix build failure during date change boundary
When the build it started on a particular date and the date changes
during the build, the final step of compressing the files fails.  This
is because the value of TODAY is not evaluated until the command is
actually executed.  This patch fixes that by make sure that the value of
variable TODAY is evaluated before the build process starts.