Don't set password for user 'fbx'
The 'fbx' user will have password disabled by default.  This is only a
problem for users who want to login on console before Plinth can create
an administrative user during first boot.  This only happens when Plinth
is not accessible due to various reasons.  While we continue to fix
problems that will make Plinth a reliable option, our fallback option is
to use a password setting tool on the downloaded image.

The password setting tool will loop mount the image and set a password
inside the image for the 'fbx' user.  This is, for now, a command line
tool that requires some knowledge of UNIX command line.  However, if the
users are wishing to login to shell as 'fbx' user, so is that process.

Essentially, in the normal flow, users won't need 'fbx' user and will
use the first boot user create by Plinth.  When Plinth is not accessible
and debugging needs to be done, users will run the password setting tool
on the image and then use the image.

TODO: Remove the 'fbx' user altogether.  Currently this is not done
because the password setting tool can't create users and wants to have
minimal dependencies to work from many OSes/distributions.