raspberry2: Have NetworkManager provide stable MAC
With Debian Stretch on Raspberry Pi 2, the kernel is not reading the
hardware MAC address, so it uses a random one. This causes the IP
address to change on every boot.

As a workaround, have NetworkManager provide a stable MAC for the
ethernet device. It is generated by hashing the device name and a
private key, so it is unique for each box.