Plant CI proof-of-concept
Only one job for the RaspberryPi target is setup. Jobs for other
targets - each with their minimal set of dependencies to install - may
be added.

Continuous integration may ease contributions to the project.

Continuous integration configuration in the repo may ease removal from
the README of the list of dependencies to install.

This CI proof-of-concept relies on Travis for offering a VM with root
priviledes, and GitHub for triggering Travis. An alternative CI
solution not bound to Travis / GitHub would be beneficial.